+++ my_extension/jupyter-extensions/jupyterlab_bigquery/src/details_panel/service/list_dataset_details.ts
