+++ my_extension/jupyter-extensions/jupyterlab_bigquery/src/list_items_panel/list_tree_item_widget.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { stylesheet } from 'typestyle';
 import { JupyterFrontEnd } from '@jupyterlab/application';
 
-import { ListProjectsService, Projects } from '../service/list_items';
+import { ListProjectsService, Projects } from './service/list_items';
 import { ListProjectItem } from './list_tree_item';
 import { WidgetManager } from '../widget_manager';
